requirements.txt
altgraph==0.17.4
black==24.4.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.2.2
pycparser==2.22
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pywin32-ctypes==0.2.2
tomli==2.0.1
typing_extensions==4.11.0
